Integrations tests, dry-run and tracking counts of changed demographic values

+++ app/models/grda_warehouse/tasks/client_cleanup.rb
@@ -7,13 +7,14 @@ module GrdaWarehouse::Tasks
     include ArelHelper
     require 'ruby-progressbar'
     attr_accessor :logger, :send_notifications, :notifier_config
-    def initialize(max_allowed=200, bogus_notifier=false, changed_client_date: 2.weeks.ago.to_date, debug: false)
+    def initialize(max_allowed=200, bogus_notifier=false, changed_client_date: 2.weeks.ago.to_date, debug: false, dry_run: false)
       @max_allowed = max_allowed
       setup_notifier('Client Cleanup')
       self.logger = Rails.logger
       @debug = debug
       @soft_delete_date = Time.now
       @changed_client_date = changed_client_date
+      @dry_run = dry_run
     end
     def run!
       remove_unused_warehouse_clients_processed()
@@ -41,7 +42,9 @@ module GrdaWarehouse::Tasks
     end
 
     def rebuild_service_history_for_incorrect_clients
-      adder = GrdaWarehouse::Tasks::ServiceHistory::Add.new
+      if ! @dry_run
+        adder = GrdaWarehouse::Tasks::ServiceHistory::Add.new
+      end
       debug_log "Rebuilding service history for #{adder.clients_needing_update_count} clients"
       adder.run!
     end
@@ -64,26 +67,28 @@ module GrdaWarehouse::Tasks
       if un_enrolled_clients.any?
         deleted_at = Time.now
         debug_log "Removing #{un_enrolled_clients.size} un enrolled source clients and associated records.  Setting DateDeleted: #{deleted_at}"
-        GrdaWarehouse::WarehouseClient.where(source_id: un_enrolled_clients).update_all(deleted_at: deleted_at)
-        GrdaWarehouse::Hud::Exit.joins(:direct_client).
-          where(Client: {id: un_enrolled_clients}).
-          update_all(DateDeleted: deleted_at)
-        GrdaWarehouse::Hud::EnrollmentCoc.joins(:direct_client).
-          where(Client: {id: un_enrolled_clients}).
-          update_all(DateDeleted: deleted_at)
-        GrdaWarehouse::Hud::Disability.joins(:direct_client).
-          where(Client: {id: un_enrolled_clients}).
-          update_all(DateDeleted: deleted_at)
-        GrdaWarehouse::Hud::HealthAndDv.joins(:direct_client).
-          where(Client: {id: un_enrolled_clients}).
-          update_all(DateDeleted: deleted_at)
-        GrdaWarehouse::Hud::IncomeBenefit.joins(:direct_client).
-          where(Client: {id: un_enrolled_clients}).
-          update_all(DateDeleted: deleted_at)
-        GrdaWarehouse::Hud::EmploymentEducation.joins(:direct_client).
-          where(Client: {id: un_enrolled_clients}).
-          update_all(DateDeleted: deleted_at)
-        GrdaWarehouse::Hud::Client.where(id: un_enrolled_clients).update_all(DateDeleted: deleted_at)
+        if ! @dry_run
+          GrdaWarehouse::WarehouseClient.where(source_id: un_enrolled_clients).update_all(deleted_at: deleted_at)
+          GrdaWarehouse::Hud::Exit.joins(:direct_client).
+            where(Client: {id: un_enrolled_clients}).
+            update_all(DateDeleted: deleted_at)
+          GrdaWarehouse::Hud::EnrollmentCoc.joins(:direct_client).
+            where(Client: {id: un_enrolled_clients}).
+            update_all(DateDeleted: deleted_at)
+          GrdaWarehouse::Hud::Disability.joins(:direct_client).
+            where(Client: {id: un_enrolled_clients}).
+            update_all(DateDeleted: deleted_at)
+          GrdaWarehouse::Hud::HealthAndDv.joins(:direct_client).
+            where(Client: {id: un_enrolled_clients}).
+            update_all(DateDeleted: deleted_at)
+          GrdaWarehouse::Hud::IncomeBenefit.joins(:direct_client).
+            where(Client: {id: un_enrolled_clients}).
+            update_all(DateDeleted: deleted_at)
+          GrdaWarehouse::Hud::EmploymentEducation.joins(:direct_client).
+            where(Client: {id: un_enrolled_clients}).
+            update_all(DateDeleted: deleted_at)
+          GrdaWarehouse::Hud::Client.where(id: un_enrolled_clients).update_all(DateDeleted: deleted_at)
+        end
       end
     end
 
@@ -186,6 +191,11 @@ module GrdaWarehouse::Tasks
     def update_client_demographics_based_on_sources
       batch_size = 1000
       processed = 0
+      changed = {
+        dobs: 0,
+        genders: 0,
+        veteran_statuses: 0,
+      }
       munge_clients = clients_to_munge
       client_source = GrdaWarehouse::Hud::Client
       total_clients = munge_clients.size
@@ -202,7 +212,8 @@ module GrdaWarehouse::Tasks
               Hash[client_columns.keys.zip(row)]
             end
           dest_attr = dest.attributes.with_indifferent_access.slice(*client_columns.keys)
-          dest_attr = choose_attributes_from_sources(dest_attr, source_clients)
+          original_attributes = dest_attr.dup
+          choose_attributes_from_sources(dest_attr, source_clients)
 
           # invalidate client if DOB has changed
           if dest.DOB != dest_attr[:DOB]
@@ -211,8 +222,16 @@ module GrdaWarehouse::Tasks
           end
           # We can speed this up if we want later.  If there's only one source client and the 
           # updated dates match, there's no need to update the destination
-          dest.update(dest_attr)
+          dest.update(dest_attr) unless @dry_run
+          changed[:dobs] += 1 if original_attributes[:DOB] != dest_attr[:DOB]
+          changed[:genders] += 1 if original_attributes[:Gender] != dest_attr[:Gender]
+          changed[:veteran_statuses] += 1 if original_attributes[:VeteranStatus] != dest_attr[:VeteranStatus]
           progress.progress += 1
+        end
+        if @debug
+          logger.debug '=========== Changed Counts ============'
+          logger.debug changed.inspect
+          logger.debug '=========== End Changed Counts ============'
         end
         processed += batch_size
         logger.info "Updated demographics for #{processed} destination clients"
@@ -265,7 +284,9 @@ module GrdaWarehouse::Tasks
           return
         end
         debug_log "Removing service history for #{non_existant_client_ids.count} clients who no longer have client records"
-        GrdaWarehouse::ServiceHistory.where(client_id: non_existant_client_ids).delete_all
+        if ! @dry_run
+          GrdaWarehouse::ServiceHistory.where(client_id: non_existant_client_ids).delete_all
+        end
       end
     end
 
@@ -278,36 +299,46 @@ module GrdaWarehouse::Tasks
         return
       end
       logger.info "Deleting Service History for #{@clients.size} clients comprising #{sh_size} records"
-      GrdaWarehouse::ServiceHistory.where(client_id: @clients).delete_all
+      if ! @dry_run
+        GrdaWarehouse::ServiceHistory.where(client_id: @clients).delete_all
+      end
     end
 
     private def clean_warehouse_clients_processed
       return unless @clients.any?
-      GrdaWarehouse::WarehouseClientsProcessed.where(client_id: @clients).delete_all
+      if ! @dry_run
+        GrdaWarehouse::WarehouseClientsProcessed.where(client_id: @clients).delete_all
+      end
     end
 
     def remove_unused_warehouse_clients_processed
       processed_ids = GrdaWarehouse::WarehouseClientsProcessed.pluck(:client_id)
       destination_client_ids = GrdaWarehouse::Hud::Client.destination.pluck(:id)
       to_remove = processed_ids - destination_client_ids
-      if to_remove.any?
+      if to_remove.any? && ! @dry_run
         GrdaWarehouse::WarehouseClientsProcessed.where(client_id: to_remove).delete_all
       end
     end
 
     private def clean_warehouse_clients
       return unless @clients.any?
-      GrdaWarehouse::WarehouseClient.where(destination_id: @clients).update_all(deleted_at: @soft_delete_date)
+      if ! @dry_run
+        GrdaWarehouse::WarehouseClient.where(destination_id: @clients).update_all(deleted_at: @soft_delete_date)
+      end
     end
 
     private def clean_hmis_clients
       return unless @clients.any?
-      GrdaWarehouse::HmisClient.where(client_id: @clients).delete_all
+      if ! @dry_run
+        GrdaWarehouse::HmisClient.where(client_id: @clients).delete_all
+      end
     end
 
     private def clean_destination_clients
       return unless @clients.any?
-      GrdaWarehouse::Hud::Client.where(id: @clients).update_all(DateDeleted: @soft_delete_date)
+      if ! @dry_run
+        GrdaWarehouse::Hud::Client.where(id: @clients).update_all(DateDeleted: @soft_delete_date)
+      end
     end
 
     def fix_incorrect_ages_in_service_history
@@ -335,10 +366,10 @@ module GrdaWarehouse::Tasks
       logger.info msg
       @notifier.ping msg if @send_notifications
       # Only invalidate clients if the age is wrong, if it's less than zero but hasn't changed, this is just wasted effort
-      GrdaWarehouse::Hud::Client.where(id: invalidate_clients.to_a).
-        map do |client|
-          client.invalidate_service_history
-        end
+      if ! @dry_run
+        GrdaWarehouse::Hud::Client.where(id: invalidate_clients.to_a).
+          each(&:invalidate_service_history)
+      end
     end
 
     def add_missing_ages_to_service_history
@@ -350,15 +381,17 @@ module GrdaWarehouse::Tasks
       if to_fix.size > 100
         @notifier.ping "Found #{to_fix.size} clients with dates of birth and service histories missing those dates.  This should not be happening.  \n\nLogical reasons include: a new import brought along a client DOB where we didn't have one before, but also had changes to enrollment, exit or services." if @send_notifications
       end
-      to_fix.each do |client_id|
-        client = GrdaWarehouse::Hud::Client.find(client_id)
-        GrdaWarehouse::Hud::Enrollment.where(
-          id: client.service_history.
-            where(age: nil).
-            joins(:enrollment).
-            select(e_t[:id].as('id').to_sql)
-        ).update_all(processed_hash: nil)
-        client.invalidate_service_history   
+      if ! @dry_run
+        to_fix.each do |client_id|
+          client = GrdaWarehouse::Hud::Client.find(client_id)
+          GrdaWarehouse::Hud::Enrollment.where(
+            id: client.service_history.
+              where(age: nil).
+              joins(:enrollment).
+              select(e_t[:id].as('id').to_sql)
+          ).update_all(processed_hash: nil)
+          client.invalidate_service_history   
+        end
       end
     end
